refactor common_fused_moe.py
